backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.2
djoser==2.2.0
webcolors==1.13
psycopg2-binary==2.9.7
Pillow==10.0.0
PyYAML==6.0
python-dotenv==1.0.0
gunicorn==20.1.0
sorl-thumbnail==12.9.0
django-import-export==3.2.0
reportlab==4.0.4
